project/backend/requirements.txt
# FastAPI Backend Dependencies
# ============================

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.13.0
# PostgreSQL drivers removed - using SQLite for Railway deployment
# asyncpg==0.29.0
# psycopg2-binary==2.9.9
aiosqlite==0.19.0

# Redis & Task Queue
redis==5.0.1
celery==5.3.4

# Vector Store
chromadb==0.4.18

# Google Gemini
google-generativeai==0.3.2

# GitHub Integration
PyGithub==2.1.1
httpx==0.25.2
aiohttp==3.9.1

# Document Processing
PyMuPDF==1.24.14
python-docx==1.1.0
markdown==3.5.1

# Data Processing
pydantic==2.10.5
pydantic-settings==2.7.1
email-validator==2.1.0
python-dotenv==1.0.0
numpy==1.26.2

# Security
cryptography==41.0.7

# Web Scraping (for JD extraction)
# playwright==1.40.0  # Disabled - requires browser binaries, causes build timeout
beautifulsoup4==4.12.3
lxml==5.3.0

# Utilities
tenacity==8.2.3
structlog==23.2.0
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
mypy==1.7.1
